Move ECSScene into engine
