+++ Greet-core/src/Greet.h
@@ -21,6 +21,7 @@
 #include <ecs/Config.h>
 #include <ecs/ECSComponents.h>
 #include <ecs/ECSManager.h>
+#include <ecs/ECSScene.h>
 #include <ecs/Entity.h>
 #include <ecs/EntityContainer.h>
 #include <ecs/EntitySet.h>
